DiscountedPrice as interface #1053
