+++ commercetools-models/src/main/java/io/sphere/sdk/productdiscounts/DiscountedPrice.java
@@ -1,30 +1,17 @@
 package io.sphere.sdk.productdiscounts;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.sphere.sdk.models.Base;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.sphere.sdk.models.Reference;
 
 import javax.money.MonetaryAmount;
 
-public final class DiscountedPrice extends Base {
-    private final MonetaryAmount value;
-    private final Reference<ProductDiscount> discount;
+@JsonDeserialize(as = DiscountedPriceImpl.class)
+public interface DiscountedPrice {
+    MonetaryAmount getValue();
 
-    @JsonCreator
-    DiscountedPrice(final MonetaryAmount value, final Reference<ProductDiscount> discount) {
-        this.value = value;
-        this.discount = discount;
-    }
+    Reference<ProductDiscount> getDiscount();
 
-    public MonetaryAmount getValue() {
-        return value;
-    }
-
-    public Reference<ProductDiscount> getDiscount() {
-        return discount;
-    }
-
-    public static DiscountedPrice of(final MonetaryAmount value, final Reference<ProductDiscount> discount) {
-        return new DiscountedPrice(value, discount);
+    static DiscountedPrice of(final MonetaryAmount value, final Reference<ProductDiscount> discount) {
+        return new DiscountedPriceImpl(value, discount);
     }
 }
